requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pydantic
python-multipart
python-jose
argon2-cffi
aiofiles